Scaffold types
Sets up core types and reducers so that the UI can hook into the actions and reducers.  Also adds a placeholder in the `optics.ts` for the work of calculating the reflections. This will be the core of this project. This diff makes no functional changes and is a placeholder for thinking and architecting.