Add MirrorLab and Observer
Adds React components for mirror lab which serves as the stage for interactions and an Observer component which represents the camera or person viewing the scene.

Note: This adds a property to Observer, isMovable, that is likely to change in the future. It's likely it will become two props for moving in the x and y dimensions but for ease of styling and getting started this makes it one prop.

This change also adds dependencies- Lucide is a library of icons. It has tree shaking so only the icons that are actually used will be included. It also adds clsx which is a tiny library that makes it easier to conditionally apply classnames.

Finally this change removes a bunch of default svgs beacuse they're not needed but also so it's harder to pinpoint that this is a nextjs app (which if a server is ever run makes it slightly more secure)