Add Control panel for observer
As of this commit the control panel shows the static information about the observer but soon (!) it will be interactive for setting properties on the observer.

This change also includes a bunch of smaller refactors and changes like changing location -> position to better align with what the domain langauge seems like.

Another change in this commit is moving the creation of the reducer and store. It's now possible to get an OpticsLab component without having it wrapped in the store. This is needed because the editor needs full control over the lab and so needs to share a reducer and state.

I imagine there will be one more version of OpticsLab that will need to instantiate its own reducer/state: I am thinking there will be a share button in the editor that will base64 encode all of the data for driving an OpticsLab so I imagine there will be something that  decodes that data and feeds it to an OpticsLab- that will probably need to instantiate its own store too.