Add Vertical Mirrors
This introduces vertical mirrors to the playground! So far mirrors aren't movable but they exist and are shown in both the control panel and in playground.

Getting here required lots of small refactors. Of note IconCard was a concept that I needed again so it moved from the Observer's card to being a more general concept.