Fix distance to mirror
This fixes up a silly mistake where the distance to the mirror wasn't quite right because, despite the mirror being a single photon wide in the simulation, it is in fact 6 pixels wide in the simulation. This nudges things by the mirror width so that everything lines up and looks shiny.