Make objects on grid stand out more
Makes objects on the grid stand out more by putting a tint on top of the dot grid pattern. This also has the effect of making the design feel more "graph paper" like. My hope is that this design feel playful and low stakes and by mimicking a real-world pritive it also feels familiar.

This helps the objects stand out more and hopefully make it more obvious that picking up the objects is how you're meant to interact with the lab