Refactor Observable Color
This refactors observable color to be part of the identity of the observable object. This is important because each added observable object should have its own color.

This also adds a color palette for observable objects. This palette was picked to avoid colors that are already in use. Some of the colors, however, are pretty close together and will definitely be hard for folks with color blindness to read. If this prototype were to continue it would be interesting to think about pattern as a second dimension on the observable objects so that they can read "read" more easily. As-is this is ok to move the prototype along