Add Observable Objects
This changes lets an editor add up to the simulation limit of observable objects (currently 5). The objects are added at a random place on the canvas and with the next available color.

There's also infrastructure in this commit for an add button which comes with some nice touches to let the user know when they can click the button and when they do click it that it's been clicked.