Make editor feel more app-like
This moves around the editor so that it feels more app-like. I'm taking design inspiration here from Apple's app (Pages, Keynote, Numbers) and from Sketch. The control pane lives on the right side and the canvas lives on the left.

This change will let me resize the canvas to different sizes and have it feel natural within the editor. It also buys more room for controls- which is good because there will be more controls in the future. It's easy to imagine that someday there will other optics-primitives like curved mirrors, lenses, point light sources and directional lights. This gives space for them all to fit in in a way that feels good.

Currently in the design there is a button that says presets but I imagine that is going to grow into a full top bar with controls that apply to the whole simulation like "share" and "preview as learner" and "load preset".