Add multiple mirrors
This adds multiple mirrors.

There is also an unusual step here of negating the use of React's keys on reflections and rays. These should NOT be cached because they are re-calculated on every move and new draw. This is a bit of a hack and given more time I'd think about a more detailed solution